Debug, fix err

+++ controllers/mylibrary.controllers.js
@@ -6,9 +6,10 @@ const Library = require("../models/library.js");
 
 
 router.get("/", isSignedIn, async (req, res) => {
-    const mylibrary = await Library.find({ user: req.session.user._id }).populate("book")
-    res.render("mylibrary/index.ejs", { mylibrary })
+    const mylibrary = await Library.find({ user: req.session.user._id }).populate("book");
+    res.render("mylibrary/index.ejs", { mylibrary });
 });
+
 
 router.post("/:bookId", isSignedIn, async (req, res) => {
 
@@ -17,18 +18,18 @@ router.post("/:bookId", isSignedIn, async (req, res) => {
         const existingBook = await Library.findOne({
             user: req.session.user._id,
             book: req.params.bookId,
-        })
+        });
 
         if (existingBook) {
-            return res.send("Book already exists in your library.")
+            return res.send("Book already exists in your library.");
         }
 
         await Library.create({
             user: req.session.user._id,
             book: req.params.bookId,
-        })
+        });
 
-        res.redirect("/mylibrary")
+        res.redirect("/mylibrary");
     }
 
     catch (err) {
@@ -36,20 +37,29 @@ router.post("/:bookId", isSignedIn, async (req, res) => {
     }
 });
 
-router.get("/:id/edit", isSignedIn, async (req, res) => {
-    const myBookEdit = await Library.findById(req.params.id)
-        .populate({
-            path: "book",
-            populate: [
-                { path: "authors" },
-                { path: "genres" }
-            ]
-        })
 
-    res.render("mylibrary/edit-my-book.ejs", { myBookEdit })
+router.get("/:id/edit", isSignedIn, async (req, res) => {
+
+    const library = await Library.findById(req.params.id);
+
+    if (library.user.toString() !== req.session.user._id.toString()) {
+        return res.send("Unauthorized.");
+    }
+
+    const myBookEdit = await Library.findById(req.params.id).populate({ path: "book", populate: [{ path: "authors" }, { path: "genres" }] });
+
+    res.render("mylibrary/edit-my-book.ejs", { myBookEdit });
 });
 
+
 router.put("/:id/edit", isSignedIn, async (req, res) => {
+
+    const library = await Library.findById(req.params.id);
+
+    if (library.user.toString() !== req.session.user._id.toString()) {
+        return res.send("Unauthorized.");
+    }
+
     await Library.findByIdAndUpdate(req.params.id, {
         status: req.body.status,
         dateStarted: req.body.dateStarted,
@@ -60,27 +70,58 @@ router.put("/:id/edit", isSignedIn, async (req, res) => {
     res.redirect("/mylibrary");
 });
 
+
 router.delete("/:id", isSignedIn, async (req, res) => {
-    await Library.findByIdAndDelete(req.params.id)
-    res.redirect("/mylibrary")
+
+    const library = await Library.findById(req.params.id);
+
+    if (library.user.toString() !== req.session.user._id.toString()) {
+        return res.send("Unauthorized.");
+    }
+
+    await Library.findByIdAndDelete(req.params.id);
+
+    res.redirect("/mylibrary");
 });
+
 
 router.post("/:id/favorite", isSignedIn, async (req, res) => {
-    const library = await Library.findById(req.params.id)
+
+    const library = await Library.findById(req.params.id);
+
+    if (library.user.toString() !== req.session.user._id.toString()) {
+        return res.send("Unauthorized.");
+    }
+
     library.isFavorited = !library.isFavorited
 
-    await library.save()
+    await library.save();
 
-    res.redirect("/mylibrary")
+    res.redirect("/mylibrary");
 });
 
+
 router.post("/:id/finish", isSignedIn, async (req, res) => {
-    const library = await Library.findById(req.params.id)
-    library.isFinished = !library.isFinished
 
-    await library.save()
+    const library = await Library.findById(req.params.id);
 
-    res.redirect("/mylibrary")
+    if (library.user.toString() !== req.session.user._id.toString()) {
+        return res.send("Unauthorized.");
+    }
+
+    if (library.status === "Finished") {
+        library.status = "Reading"
+        library.dateFinished = null
+    }
+
+    else {
+        library.status = "Finished"
+        library.dateFinished = new Date();
+    }
+
+    await library.save();
+
+    res.redirect("/mylibrary");
 });
